# backend/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "VinFast API Documentation",
    "version": "1.0.0",
    "description": "\n        ## 🚗 VinFast API Documentation\n\n        API documentation cho website cửa hàng ô tô VinFast - Hệ thống quản lý bán hàng và dịch vụ ô tô điện.\n\n        ### 🔐 Authentication\n        API sử dụng JWT (JSON Web Token) để xác thực. Để sử dụng các endpoint được bảo vệ:\n        1. Đăng nhập qua `/api/auth/login` để lấy token\n        2. Thêm token vào header: `Authorization: Bearer <token>`\n\n        ### 📱 Features\n        - **Authentication & Authorization**: Đăng ký, đăng nhập, quản lý người dùng\n        - **Car Management**: Quản lý xe, tìm kiếm, lọc theo loại\n        - **Order Management**: Đặt hàng, theo dõi đơn hàng, thanh toán\n        - **Test Drive**: Đăng ký lái thử, quản lý lịch hẹn\n        - **Contact**: Hệ thống liên hệ tư vấn\n        - **News**: Quản lý tin tức, bài viết\n        - **User Management**: Quản lý người dùng (Admin)\n\n        ### 🌐 Base URL\n        - Development: `http://localhost:8000/api`\n        - Production: `https://api.vinfast.com/api`\n      ",
    "contact": {
      "name": "VinFast Development Team",
      "email": "dev@vinfast.com",
      "url": "https://vinfast.vn"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token để xác thực. Format: Bearer <token>"
      }
    },
    "schemas": {
...
      "Car": {
        "type": "object",
        "required": [
          "name",
          "type",
          "price",
          "year",
          "description"
        ],
        "properties": {
          "_id": {
            "type": "string",
            "description": "ID xe"
          },
          "name": {
            "type": "string",
            "description": "Tên xe",
            "example": "VinFast VF 8"
          },
          "slug": {
            "type": "string",
            "description": "Slug URL",
            "example": "vinfast-vf-8"
          },
          "type": {
            "type": "string",
            "enum": [
              "sedan",
              "suv",
              "electric"
            ],
            "description": "Loại xe",
            "example": "electric"
          },
          "price": {
            "type": "number",
            "description": "Giá xe (VND)",
            "example": 1200000000
          },
          "year": {
            "type": "number",
            "description": "Năm sản xuất",
            "example": 2024
          },
          "description": {
            "type": "string",
            "description": "Mô tả xe"
          },
          "images": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "url": {
                  "type": "string"
                },
                "alt": {
                  "type": "string"
                },
                "isPrimary": {
                  "type": "boolean"
                }
              }
            }
          },
          "specifications": {
            "type": "object",
            "properties": {
              "engine": {
                "type": "string"
              },
              "power": {
                "type": "string"
              },
              "transmission": {
                "type": "string"
              },
              "fuelType": {
                "type": "string"
              },
              "seating": {
                "type": "number"
              }
            }
          },
          "isElectric": {
            "type": "boolean",
            "description": "Xe điện",
            "example": true
          },
          "isFeatured": {
            "type": "boolean",
            "description": "Xe nổi bật",
            "example": true
          },
          "stock": {
            "type": "number",
            "description": "Số lượng tồn kho",
            "example": 10
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "message": {
            "type": "string",
            "description": "Thông báo lỗi"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "description": "Chi tiết lỗi validation"
          }
        }
      },
      "Success": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "description": "Thông báo thành công",
            "example": "Thành công"
          },
          "data": {
            "type": "object",
            "description": "Dữ liệu trả về"
          },
          "count": {
            "type": "number",
            "description": "Số lượng items trong trang hiện tại",
            "example": 10
          },
          "total": {
            "type": "number",
            "description": "Tổng số items",
            "example": 100
          },
          "pagination": {
            "type": "object",
            "properties": {
              "next": {
                "type": "object",
                "properties": {
                  "page": {
                    "type": "number",
                    "example": 2
                  },
                  "limit": {
                    "type": "number",
                    "example": 10
                  }
                }
              },
              "prev": {
                "type": "object",
                "properties": {
                  "page": {
                    "type": "number",
                    "example": 1
                  },
                  "limit": {
                    "type": "number",
                    "example": 10
                  }
                }
              }
            }
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/cars": {
      "get": {
        "summary": "Lấy danh sách xe",
        "tags": [
          "Cars"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1
            },
            "description": "Số trang"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50
            },
            "description": "Số lượng xe mỗi trang"
          },
          {
            "in": "query",
            "name": "type",
            "schema": {
              "type": "string",
              "enum": [
                "sedan",
                "suv",
                "electric"
              ]
            },
            "description": "Loại xe"
          },
          {
            "in": "query",
            "name": "minPrice",
            "schema": {
              "type": "number"
            },
            "description": "Giá tối thiểu"
          },
          {
            "in": "query",
            "name": "maxPrice",
            "schema": {
              "type": "number"
            },
            "description": "Giá tối đa"
          },
          {
            "in": "query",
            "name": "year",
            "schema": {
              "type": "integer"
            },
            "description": "Năm sản xuất"
          }
        ],
        "responses": {
          "200": {
            "description": "Danh sách xe",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Success"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Car"
                          }
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Tạo xe mới (Admin only)",
        "tags": [
          "Cars"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "type",
                  "price",
                  "year",
                  "description",
                  "specifications"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 100,
                    "example": "VinFast VF 9"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "sedan",
                      "suv",
                      "electric"
                    ],
                    "example": "electric"
                  },
                  "price": {
                    "type": "number",
                    "minimum": 0,
                    "example": 1500000000
                  },
                  "year": {
                    "type": "integer",
                    "minimum": 2020,
                    "example": 2024
                  },
                  "description": {
                    "type": "string",
                    "minLength": 10,
                    "maxLength": 1000,
                    "example": "VinFast VF 9 là mẫu SUV điện cao cấp..."
                  },
                  "specifications": {
                    "type": "object",
                    "required": [
                      "engine",
                      "power",
                      "seating"
                    ],
                    "properties": {
                      "engine": {
                        "type": "string",
                        "example": "Động cơ điện"
                      },
                      "power": {
                        "type": "string",
                        "example": "408 HP"
                      },
                      "seating": {
                        "type": "integer",
                        "minimum": 2,
                        "maximum": 9,
                        "example": 7
                      }
                    }
                  },
                  "images": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "url": {
                          "type": "string"
                        },
                        "alt": {
                          "type": "string"
                        },
                        "isPrimary": {
                          "type": "boolean"
                        }
                      }
                    }
                  },
                  "isFeatured": {
                    "type": "boolean",
                    "example": true
                  },
                  "stock": {
                    "type": "integer",
                    "minimum": 0,
                    "example": 10
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Tạo xe thành công",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Success"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Car"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Dữ liệu không hợp lệ",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Không có quyền truy cập",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Không có quyền admin",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/cars/{id}": {
      "get": {
        "summary": "Lấy chi tiết xe",
        "tags": [
          "Cars"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID của xe"
          }
        ],
        "responses": {
          "200": {
            "description": "Chi tiết xe",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Success"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Car"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy xe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Cập nhật xe (Admin only)",
        "tags": [
          "Cars"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID của xe"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 2,
                    "maxLength": 100
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "sedan",
                      "suv",
                      "electric"
                    ]
                  },
                  "price": {
                    "type": "number",
                    "minimum": 0
                  },
                  "year": {
                    "type": "integer",
                    "minimum": 2020
                  },
                  "description": {
                    "type": "string"
                  },
                  "isFeatured": {
                    "type": "boolean"
                  },
                  "isActive": {
                    "type": "boolean"
                  },
                  "stock": {
                    "type": "integer",
                    "minimum": 0
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cập nhật xe thành công",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/Success"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/Car"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy xe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Xóa xe (Admin only)",
        "tags": [
          "Cars"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "ID của xe"
          }
        ],
        "responses": {
          "200": {
            "description": "Xóa xe thành công",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          },
          "404": {
            "description": "Không tìm thấy xe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}